Completed challenge six of flex in foundations
Added display flex to body and changed the direction
to column, added padding to the buttons, added flex
to header, content, footer classes and the ul element.
Set the margin-right of left-links to auto to push
the left-links and right-links to either side. Also,
centered the elements inside the content class after
changing the flex-direction to column.